backend/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyPDF2==3.0.1
Werkzeug==3.1.3
Pillow==10.4.0
pillow-heif==0.17.0
